tests/wpt/tests/speculation-rules/prerender/windowclient-navigate-to-same-origin-url-on-iframe.https.html

<!DOCTYPE html>
<title>WindowClient.navigate() to same-origin url in a prerendered iframe</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/utils.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

const PAGE_URL = 'resources/windowclient-navigate-on-iframe.html';
const WORKER_URL = 'resources/windowclient-navigate-worker.js';
const SCOPE = 'resources/';
const SAME_ORIGIN_DESTINATION =
    get_host_info()['HTTPS_ORIGIN'] + base_path() + 'resources/empty.html';

promise_test(async t => {
  const uid = token();

  const registration = await service_worker_unregister_and_register(
      t, `${WORKER_URL}?uid=${uid}`, SCOPE);
  t.add_cleanup(() => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');

  const bc = new PrerenderChannel('test-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // PAGE_URL starts a prerender of a page that makes an iframe, then asks the
  // service worker to navigate the iframe to `navigationUrl` via
  // `WindowClient.navigate(url)`.
  window.open(
      `${PAGE_URL}?navigationUrl=${SAME_ORIGIN_DESTINATION}&uid=${uid}`,
      '_blank',
      'noopener');

  const navigationResult = await gotMessage;
  assert_equals(navigationResult, 'navigate() succeeded',
      'should succeed to finish navigation test');

  // Send a close signal to PrerenderEventCollector on the prerendered page.
  new PrerenderChannel('close', uid).postMessage('');
}, 'WindowClient.navigate() to a same-origin URL on a prerendered iframe ' +
   'should succeed');
</script>
